requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-tokenapi==0.2.3
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
httpie==0.9.2
psycopg2==2.6.1
Pygments==2.0.2
requests==2.8.1
static3==0.6.1
wheel==0.24.0
